# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sales & Orders API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Browse/search product catalog
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: inStock
          in: query
          schema:
            type: ["boolean", "null"]
        - name: customerId
          in: query
          schema:
            type: ["integer", "null"]
          description: When set, sorts frequent items for this customer first
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update catalog product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
  /products/{id}/sales:
    get:
      operationId: getProductSales
      tags: [products]
      summary: Product sales summary for a date range
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Product sales summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductSalesSummary"

  # ── Orders ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, sku, name, unitPrice, inventoryQty, inStock]
      properties:
        id:
          type: integer
        sku:
          type: string
        name:
          type: string
        unitPrice:
          type: number
        inventoryQty:
          type: integer
        inStock:
          type: boolean
        etaDate:
          type: ["string", "null"]
          format: date
        imageUrl:
          type: ["string", "null"]
        packSize:
          type: ["string", "null"]
        orderFrequency:
          type: ["integer", "null"]
          description: How often this customer orders this product
    ProductDetail:
      type: object
      required: [id, sku, name, unitPrice, inventoryQty, inStock]
      properties:
        id:
          type: integer
        sku:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        unitPrice:
          type: number
        inventoryQty:
          type: integer
        inStock:
          type: boolean
        etaDate:
          type: ["string", "null"]
          format: date
        imageUrl:
          type: ["string", "null"]
        packSize:
          type: ["string", "null"]
        certifications:
          type: array
          items:
            type: string
        brochureUrl:
          type: ["string", "null"]
        infoSheetUrl:
          type: ["string", "null"]
        activePromotion:
          nullable: true
          oneOf:
            - $ref: "#/components/schemas/Promotion"
            - type: "null"
    ProductUpdate:
      type: object
      properties:
        sku:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        unitPrice:
          type: number
        inventoryQty:
          type: integer
        etaDate:
          type: ["string", "null"]
          format: date
        imageUrl:
          type: ["string", "null"]
        packSize:
          type: ["string", "null"]
        certifications:
          type: array
          items:
            type: string
        brochureUrl:
          type: ["string", "null"]
        infoSheetUrl:
          type: ["string", "null"]
    ProductSalesSummary:
      type: object
      required: [productId, startDate, endDate, unitsSold, orderCount, revenue]
      properties:
        productId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        unitsSold:
          type: integer
        orderCount:
          type: integer
        revenue:
          type: number

    # ── Orders ────────────────────────────────────────────────────────────────
...
    Promotion:
      type: object
      required: [id, name, discountType, discountValue]
      properties:
        id:
          type: integer
        name:
          type: string
        discountType:
          type: string
          enum: [percent, fixed]
        discountValue:
          type: number
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        productIds:
          type: array
          items:
            type: integer